Fix Social Studio and KPI inconsistencies

+++ backend/app/schemas.py
@@ -179,7 +179,7 @@ class PostMetricIn(BaseModel):
     post_url: str = ""
     post_id: str = ""
     published_date: str | None = None
-    spend: int = 0
+    spend: float = 0
     impressions: int = 0
     reach: int = 0
     clicks: int = 0
